requirements.txt
jsonsearch==0.0.2
Pillow==9.4.0
requests==2.28.1
rich==13.0.1
selenium==3.14.0
SocksiPy_branch==1.01
threadpool==1.3.2
